Brings the branch up to upstream master (e071055f): 743 commits of
source, examples, tools and lib/kokkos (5.2.2), plus the doc content
that landed upstream since this branch forked at 09391fe8.

The doc merge could not be done by text alone: both branches convert
doc/*.txt to doc/src/*.rst independently, so 106 of the 141 conflicts
were add/add with no common ancestor.  The upstream content delta was
isolated instead by running this branch's own converter pipeline
(utils/converters: txt2rst, postconvert.py, split_chapters.py -- which
documents itself as idempotent and re-runnable) over both 09391fe8 and
upstream/master, giving a real merge base and a real "theirs" in this
branch's own style.  A per-file three-way merge against those applied
the content change and nothing else: 57 files merged clean, 64 hunks
needed a decision, of which 19 were pure rst escaping, rewrapping or a
URL trailing slash.

The 45 judged by hand were almost all this branch's deliberate repairs
meeting the unrepaired upstream text -- fixed typos, :doc:/:ref:
cross-reference restructuring, the per-topic page split, citations
collected onto Bibliography.  Those keep this branch's version.  Six
took upstream's:

  Speed_kokkos     the Trilinos link, which upstream repointed at
                   trilinos/Trilinos (this branch still had kokkos/kokkos)
  dump_image       the persp keyword documentation, which upstream
                   dropped -- src/dump_image.cpp errors out on it
  dump_image       the Section 2.2.2 anchor: both branches repaired the
                   broken one, upstream's is the more precise
  collide_modify   the partners keyword, rewritten upstream
  History_future   the stale feature request form, dropped upstream
  fix_ablate       a new paragraph on dropped decrements (merged with
                   this branch's "applied" spelling fix)

Also fixed, all found by the checks rather than by reading:

- Sun05 is a citation upstream added; it was the one label in the
  manual with no entry on the Bibliography page.
- Bird09 labels two different papers upstream, one in collide_modify
  and one in react.  Per-page anchors hid the collision; a single
  Bibliography page cannot, and collide_modify was citing react's
  paper.  Split into Bird09 and Bird09b.
- surf_collide's impulsive/softsphere equation used \textlangle and
  \textrangle, which need a LaTeX package MathJax does not load, so it
  rendered the macro names in red.  \langle and \rangle are the same
  characters.
- doc/utils/sphinx-config/conf.py came in from the other branch, which
  commits it; this branch generates it from conf.py.in and ignores it.
  The stale copy shadowed the generated one, silently disabling the
  SPARTA lexer and the _ext directives (348 warnings).
- The two branches each added a doc CI job.  Kept this branch's, which
  is the superset, and carried over the other's Python matrix, which
  exists because 3.9 and 3.11 do not resolve to the same Sphinx and do
  not report the same things.

Verified:

  sphinx-build -W --keep-going   clean (also under Sphinx 8.1.3)
  make -C doc check              134 anchors, 175 styles, no errors
  equation-check.py              24/24 match
  link-check.py                  164 dead links, byte-identical to the
                                 set on this branch before the merge --
                                 all pre-existing, none introduced here
                                 (23 orphan full-size JPGs and
                                 site-absolute /doc/*.html links)
  content audit                  every line of the upstream delta
                                 accounted for in the merged tree

+++ src/fix_ave_histo_weight.cpp
@@ -319,8 +319,8 @@ void FixAveHistoWeight::bin_one(double value)
 
 void FixAveHistoWeight::bin_vector(int n, double *values, int stride)
 {
-  int m = 0;
-  int mwt = 0;
+  bigint m = 0;
+  bigint mwt = 0;
   for (int i = 0; i < n; i++) {
     bin_one_weight(values[m],weights[mwt]);
     m += stride;
@@ -343,7 +343,7 @@ void FixAveHistoWeight::bin_particles(int attribute, int index)
   Region *region;
   if (regionflag) region = domain->regions[iregion];
 
-  int mwt = 0;
+  bigint mwt = 0;
 
   if (attribute == X) {
     if (regionflag && mixflag) {
@@ -415,8 +415,8 @@ void FixAveHistoWeight::bin_particles(double *values, int stride)
   Region *region;
   if (regionflag) region = domain->regions[iregion];
 
-  int m = 0;
-  int mwt = 0;
+  bigint m = 0;
+  bigint mwt = 0;
 
   if (regionflag && mixflag) {
     for (int i = 0; i < nlocal; i++) {
@@ -458,8 +458,8 @@ void FixAveHistoWeight::bin_grid_cells(double *values, int stride)
   Grid::ChildInfo *cinfo = grid->cinfo;
   int nglocal = grid->nlocal;
 
-  int m = 0;
-  int mwt = 0;
+  bigint m = 0;
+  bigint mwt = 0;
 
   if (groupflag) {
     for (int i = 0; i < nglocal; i++) {
